requirements.txt
pip~=23.2.1
pythonnet~=3.0.5
pywebview~=5.4
typing_extensions~=4.12.2
bottle~=0.13.2
proxy_tools~=0.1.0
pycparser~=2.22
clr_loader~=0.2.7.post0
cffi~=1.17.1
pandas~=2.2.3
numpy~=2.2.3
openpyxl~=3.1.5